+++ jdi-light-bootstrap-tests/src/main/java/io/github/com/pages/BootstrapPage.java
@@ -114,18 +114,18 @@ import io.github.com.sections.navbar.NavbarText;
 import io.github.com.sections.navbar.NavbarWithDropdown;
 import io.github.com.sections.navbar.NavbarWithInputGroupForm;
 import io.github.com.sections.navbar.ScrollSpyNav;
-import io.github.com.sections.navs.NavsAlignmentCenter;
-import io.github.com.sections.navs.NavsAlignmentEnd;
-import io.github.com.sections.navs.NavsBaseA;
-import io.github.com.sections.navs.NavsBaseLi;
-import io.github.com.sections.navs.NavsJustify;
-import io.github.com.sections.navs.NavsPills;
-import io.github.com.sections.navs.NavsPillsWithDropdown;
-import io.github.com.sections.navs.NavsTabs;
-import io.github.com.sections.navs.NavsTabsWithDropdown;
-import io.github.com.sections.navs.NavsVerticalA;
-import io.github.com.sections.navs.NavsVerticalLi;
-import io.github.com.sections.navs.NestedNav;
+import com.epam.jdi.light.ui.bootstrap.elements.complex.navs.NavsAlignmentCenter;
+import com.epam.jdi.light.ui.bootstrap.elements.complex.navs.NavsAlignmentEnd;
+import com.epam.jdi.light.ui.bootstrap.elements.complex.navs.NavsBaseA;
+import com.epam.jdi.light.ui.bootstrap.elements.complex.navs.NavsBaseLi;
+import com.epam.jdi.light.ui.bootstrap.elements.complex.navs.NavsJustify;
+import com.epam.jdi.light.ui.bootstrap.elements.complex.navs.NavsPills;
+import com.epam.jdi.light.ui.bootstrap.elements.complex.navs.NavsPillsWithDropdown;
+import com.epam.jdi.light.ui.bootstrap.elements.complex.navs.NavsTabs;
+import com.epam.jdi.light.ui.bootstrap.elements.complex.navs.NavsTabsWithDropdown;
+import com.epam.jdi.light.ui.bootstrap.elements.complex.navs.NavsVerticalA;
+import com.epam.jdi.light.ui.bootstrap.elements.complex.navs.NavsVerticalLi;
+import com.epam.jdi.light.ui.bootstrap.elements.complex.navs.NestedNav;
 import io.github.com.sections.pagination.PaginationAlignCenter;
 import io.github.com.sections.pagination.PaginationAlignEnd;
 import io.github.com.sections.pagination.PaginationIcons;
