+++ jdi-light-bootstrap/src/main/java/com/epam/jdi/light/ui/bootstrap/elements/complex/navs/NavsAlignmentCenter.java
@@ -1,12 +1,14 @@
-package io.github.com.sections.navs;
+package com.epam.jdi.light.ui.bootstrap.elements.complex.navs;
 
 import com.epam.jdi.light.elements.composite.Section;
 import com.epam.jdi.light.elements.pageobjects.annotations.locators.UI;
 import com.epam.jdi.light.ui.bootstrap.elements.complex.ListGroup;
 
-public class NavsVerticalA extends Section {
+public class NavsAlignmentCenter extends Section {
 
-    @UI("a")
+    @UI("li")
+    public ListGroup navItem;
+    @UI("li a")
     public ListGroup navItemLink;
 
 }
