+++ jdi-light-bootstrap/src/main/java/com/epam/jdi/light/ui/bootstrap/elements/complex/navs/NavsBaseA.java
@@ -1,14 +1,12 @@
-package io.github.com.sections.navs;
+package com.epam.jdi.light.ui.bootstrap.elements.complex.navs;
 
 import com.epam.jdi.light.elements.composite.Section;
 import com.epam.jdi.light.elements.pageobjects.annotations.locators.UI;
 import com.epam.jdi.light.ui.bootstrap.elements.complex.ListGroup;
 
-public class NavsTabs extends Section {
+public class NavsBaseA extends Section {
 
-    @UI("li")
-    public ListGroup navItem;
-    @UI("li a")
+    @UI("a")
     public ListGroup navItemLink;
 
 }
