+++ jdi-light-bootstrap/src/main/java/com/epam/jdi/light/ui/bootstrap/elements/complex/navs/NavsJustify.java
@@ -1,10 +1,10 @@
-package io.github.com.sections.navs;
+package com.epam.jdi.light.ui.bootstrap.elements.complex.navs;
 
 import com.epam.jdi.light.elements.composite.Section;
 import com.epam.jdi.light.elements.pageobjects.annotations.locators.UI;
 import com.epam.jdi.light.ui.bootstrap.elements.complex.ListGroup;
 
-public class NavsPills extends Section {
+public class NavsJustify extends Section {
 
     @UI("li")
     public ListGroup navItem;
