+++ jdi-light-bootstrap/src/main/java/com/epam/jdi/light/ui/bootstrap/elements/complex/navs/NavsTabsWithDropdown.java
@@ -1,4 +1,4 @@
-package io.github.com.sections.navs;
+package com.epam.jdi.light.ui.bootstrap.elements.complex.navs;
 
 import com.epam.jdi.light.elements.complex.dropdown.Dropdown;
 import com.epam.jdi.light.elements.composite.Section;
