+++ jdi-light-bootstrap/src/main/java/com/epam/jdi/light/ui/bootstrap/elements/complex/navs/NestedNav.java
@@ -1,4 +1,4 @@
-package io.github.com.sections.navs;
+package com.epam.jdi.light.ui.bootstrap.elements.complex.navs;
 
 import com.epam.jdi.light.elements.composite.Section;
 import com.epam.jdi.light.elements.pageobjects.annotations.locators.UI;
